changed file structure for resoving building errors, added cast from NumericType to NumericType
